Add files via upload
One of the main problem of our dataset is that, from our data, a netflix user cannot understand which movie or tv show to watch that is similar to the one they previously watched. The other main problem is that, when a new film or tv show is released, it will take time to see its IMDb rating. However, that would be great if we have a prediction algrithm for newly released films to predict their IMDb ratings.

Also, in order to visualize and analyze our data in a better way we thought that; data cleaning, EDA and dimensionality reduction needs to be performed.